[refactor] MonthlyReport 폴더 구조 변경
