+++ src/components/MonthlyReport/index.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import PieChartContainer from "@/components/PieChartContainer";
 import Button from "@/components/base/Button";
 import React from "react";
 import { TbReportAnalytics } from "react-icons/tb";
 import excelDownloader from "@/utils/excel";
-import ProgressContainer from "../ProgressContainer";
+import ProgressContainer from "./items/ProgressContainer";
+import PieChartContainer from "./items/PieChartContainer";
 
 const MonthlyReport = () => {
   return (
